+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Add `Signal#all` method
+
 ## 3.0.2 (2019-01-26)
 
 * Fix issue with array arguments passed to `startWith`, `endWith`, `append`, and `prepend` methods